Conflicts:
	lucene/core/src/java/org/apache/lucene/search/CachingWrapperFilter.java
	lucene/core/src/java/org/apache/lucene/search/ConjunctionScorer.java
	lucene/core/src/java/org/apache/lucene/search/ConjunctionTermScorer.java
	lucene/core/src/java/org/apache/lucene/search/ScoreCachingWrappingScorer.java
	lucene/core/src/java/org/apache/lucene/search/SloppyPhraseScorer.java
	lucene/core/src/test/org/apache/lucene/search/JustCompileSearch.java

+++ lucene/core/src/java/org/apache/lucene/search/BooleanScorer2.java
@@ -129,7 +129,7 @@ class BooleanScorer2 extends Scorer {
     }
 
     @Override
-    public float freq() throws IOException {
+    public int freq() throws IOException {
       return 1;
     }
 
@@ -318,7 +318,7 @@ class BooleanScorer2 extends Scorer {
   }
 
   @Override
-  public float freq() throws IOException {
+  public int freq() throws IOException {
     return countingSumScorer.freq();
   }
