+++ lucene/core/src/java/org/apache/lucene/search/ExactPhraseScorer.java
@@ -189,7 +189,7 @@ final class ExactPhraseScorer extends Scorer {
   }
 
   @Override
-  public float freq() {
+  public int freq() {
     return freq;
   }
